Albuminuria Staging based on ACR

Using KDIGO 2012 Clinical Practice Guideline for the Evaluation and Management of Chronic Kidney Disease Volume 3 | Issue 1 | January 2013

Usage

Albuminuria_staging_ACR(...)

# S3 method for class 'data.frame'
Albuminuria_staging_ACR(.data, ACR, ...)

# S3 method for class 'units'
Albuminuria_staging_ACR(ACR, ...)

# S3 method for class 'numeric'
Albuminuria_staging_ACR(ACR, ...)

Arguments

...

Further optional arguments

.data

(data.frame) A data.frame, optional

ACR

Albumin-to-creatinine ratio column name, or vector of units or numeric if .data is not provided

Value

Albuminuria category

Details

A1: Normal to mildly increased

A2: Moderately increased

A3: Severely increased

See https://kdigo.org/guidelines/ckd-evaluation-and-management/ for more details

Examples

df <- tibble::tibble(
  ACR = units::set_units(c(-1, NA, 1, 10, 50), "mg/g")
)

Albuminuria_staging_ACR(df, "ACR")
#> [1] No Albuminuria No Albuminuria A1             A2             A3            
#> Levels: No Albuminuria < A1 < A2 < A3

df %>%
  dplyr::mutate(GFR_level = Albuminuria_staging_ACR(ACR))
#> # A tibble: 5 × 2
#>      ACR GFR_level     
#>   [mg/g] <ord>         
#> 1     -1 No Albuminuria
#> 2     NA No Albuminuria
#> 3      1 A1            
#> 4     10 A2            
#> 5     50 A3