Albuminuria Staging based on AER

Using KDIGO 2012 Clinical Practice Guideline for the Evaluation and Management of Chronic Kidney Disease Volume 3 | Issue 1 | January 2013

Albuminuria_staging_AER(...)

# S3 method for class 'data.frame'
Albuminuria_staging_AER(.data, AER, ...)

# S3 method for class 'units'
Albuminuria_staging_AER(AER, ...)

# S3 method for class 'numeric'
Albuminuria_staging_AER(AER, ...)

Arguments

...

Further optional arguments

.data

(data.frame) A data.frame, optional

AER

Albumin excretion rate column name, or vector of units or numeric if .data is not provided

Value

Albuminuria category

Details

A1: Normal to mildly increased

A2: Moderately increased

A3: Severely increased

See https://kdigo.org/guidelines/ckd-evaluation-and-management/ for more details

Examples

df <- tibble::tibble(
  AER = units::set_units(c(-1, NA, 15, 100, 500), "mg/day")
)

Albuminuria_staging_AER(df, "AER")
#> [1] No Albuminuria No Albuminuria A1             A2             A3            
#> Levels: No Albuminuria < A1 < A2 < A3

df %>%
  dplyr::mutate(GFR_level = Albuminuria_staging_AER(AER))
#> # A tibble: 5 × 2
#>      AER GFR_level     
#>   [mg/d] <ord>         
#> 1     -1 No Albuminuria
#> 2     NA No Albuminuria
#> 3     15 A1            
#> 4    100 A2            
#> 5    500 A3