Codify AKI from Serum Creatinine and/or Urine Output

Using KDIGO Clinical Practice Guideline for Acute Kidney Injury Volume 2 | Issue 1 | March 2012

aki(...)

# S3 method for numeric
aki(SCr, bCr = NULL, units = "umol/l", na.rm = FALSE, ...)

# S3 method for units
aki(SCr, bCr = NULL, na.rm = FALSE, ...)

# S3 method for ts
aki(SCr, UO, units = "umol/l", ...)

# S3 method for default
aki(
  data,
  SCr = NULL,
  bCr = NULL,
  UO = NULL,
  aki = "aki",
  units = list(SCr = "umol/l"),
  na.rm = FALSE,
  ...
)

Arguments

...
Further optional arguments that will be passed to method.

SCr
The variable name, e.g. "cr", to be used for determining AKI based on creatinine level. A numeric vector or time series of serum creatinine values if the data argument is unused

bCr
The variable name, e.g. "baseline_cr", to be used for determining AKI based on urine output. A single numeric value of baseline creatinine if the data argument is unused

units
(character) Units of SCr and UO metric (mg/dl) or SI (umol/l) #TOFIX, consider changing to a list

na.rm
(logical) If TRUE, missing values are removed

UO
The variable name, e.g. "urine", to be used for determining AKI based on urine output. A numeric vector or time series of urine output values if the data argument is unused

data
A data.frame with the data needed for serum creatinine (SCr) and/or urine output (UO)

aki
the variable name e.g. "aki_stages" to be used for the output

Details

Provided a series of Serum Creatinine readings and/or Urine Output, aki() calculates whether or not a patient has AKI. The staging (1, 2, 3) of AKI is also calculated

Examples

aki(seq(60, 200, by = 20))

#> [1] No AKI      No AKI      AKI Stage 1 AKI Stage 2 AKI Stage 2 AKI Stage 2
#> [7] AKI Stage 3 AKI Stage 3
#> Levels: AKI Stage 1 AKI Stage 2 AKI Stage 3 No AKI

aki(SCr = seq(60, 200, by = 20), bCr = 50)

#> [1] No AKI      AKI Stage 1 AKI Stage 2 AKI Stage 2 AKI Stage 2 AKI Stage 3
#> [7] AKI Stage 3 AKI Stage 3
#> Levels: AKI Stage 1 AKI Stage 2 AKI Stage 3 No AKI