AKI Staging based on Baseline Serum Creatinine

Using KDIGO Clinical Practice Guideline for Acute Kidney Injury Volume 2 | Issue 1 | March 2012

aki_bCr(...)

# S3 method for class 'data.frame'
aki_bCr(.data, SCr, bCr, ...)

# S3 method for class 'units'
aki_bCr(SCr, bCr, ...)

# S3 method for class 'numeric'
aki_bCr(SCr, bCr, ...)

Arguments

...

Further optional arguments

.data

(data.frame) A data.frame, optional

SCr

Serum creatinine column name, or vector of units or numeric if .data is not provided

bCr

Baseline creatinine column name, or vector of units or numeric if .data is not provided

Value

(ordered factor) AKI stages

Details

AKI Stage 1: 1.5-1.9 times baseline

AKI Stage 2: 2.0-2.9 times baseline

AKI Stage 3: 3.0 times baseline

See https://kdigo.org/guidelines/acute-kidney-injury/ for more details.

Examples

aki_bCr(aki_pt_data, SCr = "SCr_", bCr = "bCr_")
#>  [1] No AKI      AKI Stage 1 AKI Stage 2 AKI Stage 2 AKI Stage 3 AKI Stage 3
#>  [7] No AKI      No AKI      No AKI      No AKI      No AKI      No AKI     
#> [13] No AKI      No AKI      No AKI      No AKI      No AKI      No AKI     
#> [19] No AKI      No AKI      No AKI      No AKI      No AKI      No AKI     
#> [25] No AKI      No AKI      No AKI     
#> Levels: No AKI < AKI Stage 1 < AKI Stage 2 < AKI Stage 3

aki_pt_data %>%
  dplyr::mutate(aki = aki_bCr(SCr = SCr_, bCr = bCr_))
#> # A tibble: 27 × 8
#>       SCr_    bCr_ pt_id_ dttm_                 UO_ aki_staging_type aki_  aki  
#>    [mg/dl] [mg/dl] <chr>  <dttm>              [ml/… <chr>            <ord> <ord>
#>  1     2       1.5 NA     NA                     NA aki_bCr          No A… No A…
#>  2     2.5     1.5 NA     NA                     NA aki_bCr          AKI … AKI …
#>  3     3       1.5 NA     NA                     NA aki_bCr          AKI … AKI …
#>  4     3.5     1.5 NA     NA                     NA aki_bCr          AKI … AKI …
#>  5     4       1.5 NA     NA                     NA aki_bCr          AKI … AKI …
#>  6     4.5     1.5 NA     NA                     NA aki_bCr          AKI … AKI …
#>  7     3.4    NA   pt1    2020-10-23 09:00:00    NA aki_SCr          No A… No A…
#>  8     3.9    NA   pt1    2020-10-25 21:00:00    NA aki_SCr          No A… No A…
#>  9     3      NA   pt1    2020-10-20 09:00:00    NA aki_SCr          AKI … No A…
#> 10     3.4    NA   pt2    2020-10-18 22:00:00    NA aki_SCr          No A… No A…
#> # ℹ 17 more rows