Codify AKI from Serum Creatinine and/or Urine Output

Using KDIGO Clinical Practice Guideline for Acute Kidney Injury Volume 2 | Issue 1 | March 2012

Usage

aki_staging(...)

# S3 method for class 'data.frame'
aki_staging(
  .data,
  SCr = NULL,
  bCr = NULL,
  UO = NULL,
  dttm = NULL,
  pt_id = NULL,
  ...
)

# S3 method for class 'units'
aki_staging(SCr = NULL, bCr = NULL, UO = NULL, dttm = NULL, pt_id = NULL, ...)

# S3 method for class 'numeric'
aki_staging(SCr = NULL, bCr = NULL, UO = NULL, dttm = NULL, pt_id = NULL, ...)

Arguments

...

Further optional arguments

.data

(data.frame) A data.frame, optional

SCr

Serum creatinine column name, or vector of units or numeric if .data is not provided

bCr

Baseline creatinine column name, or vector of units or numeric if .data is not provided

UO

Urine output column name, or vector of units or numeric if .data is not provided

dttm

DateTime column name, or vector of POSIXct if .data is not provided

pt_id

Patient ID column name, or vector of characters or factors if .data is not provided

Value

(ordered factor) AKI stages

Details

Provided a baseline creatinine, series of Serum Creatinine readings and/or Urine Output, aki_staging() calculates whether or not a patient has AKI. The staging (1, 2, 3) of AKI is returned.

When multiple columns are provided, aki_staging() will automatically calculate whether or not AKI has occurred using each KDIGO definition.

aki_bCr(): Staging of AKI based on baseline serum creatinine

aki_SCr(): Staging of AKI based on changes in serum creatinine

aki_UO(): Staging of AKI based on urine output

The most severe AKI stage is then returned.

See https://kdigo.org/guidelines/acute-kidney-injury/ for more details.

Examples

aki_staging(aki_pt_data, SCr = "SCr_", bCr = "bCr_", UO = "UO_", dttm = "dttm_", pt_id = "pt_id_")
#>  [1] No AKI      AKI Stage 1 AKI Stage 2 AKI Stage 2 AKI Stage 3 AKI Stage 3
#>  [7] No AKI      No AKI      AKI Stage 1 No AKI      No AKI      AKI Stage 1
#> [13] No AKI      No AKI      No AKI      AKI Stage 1 No AKI      AKI Stage 2
#> [19] AKI Stage 3 AKI Stage 1 AKI Stage 3 AKI Stage 2 No AKI      AKI Stage 1
#> [25] AKI Stage 3 AKI Stage 3 No AKI     
#> Levels: No AKI < AKI Stage 1 < AKI Stage 2 < AKI Stage 3

aki_pt_data %>%
  dplyr::mutate(aki = aki_staging(SCr = SCr_, bCr = bCr_, UO = UO_, dttm = dttm_, pt_id = pt_id_))
#> # A tibble: 27 × 8
#>       SCr_    bCr_ pt_id_ dttm_                 UO_ aki_staging_type aki_  aki  
#>    [mg/dl] [mg/dl] <chr>  <dttm>              [ml/… <chr>            <ord> <ord>
#>  1     2       1.5 NA     NA                     NA aki_bCr          No A… No A…
#>  2     2.5     1.5 NA     NA                     NA aki_bCr          AKI … AKI …
#>  3     3       1.5 NA     NA                     NA aki_bCr          AKI … AKI …
#>  4     3.5     1.5 NA     NA                     NA aki_bCr          AKI … AKI …
#>  5     4       1.5 NA     NA                     NA aki_bCr          AKI … AKI …
#>  6     4.5     1.5 NA     NA                     NA aki_bCr          AKI … AKI …
#>  7     3.4    NA   pt1    2020-10-23 09:00:00    NA aki_SCr          No A… No A…
#>  8     3.9    NA   pt1    2020-10-25 21:00:00    NA aki_SCr          No A… No A…
#>  9     3      NA   pt1    2020-10-20 09:00:00    NA aki_SCr          AKI … AKI …
#> 10     3.4    NA   pt2    2020-10-18 22:00:00    NA aki_SCr          No A… No A…
#> # ℹ 17 more rows