Calculate age from DOB

Calculate age from DOB

dob2age(dob, age_on = lubridate::today(), fun = NULL, units = "years", ...)

Arguments

dob
The date or vector of dates representing date(s) of birth.

age_on
(Date) The date on which age is to be calculated. Defaults to today.

fun
(function) The function to be applied to the age, e.g. floor. Defaults to NULL.

units
(character) The units to measure age in, e.g. "years". Only used if fun is specified. Defaults to "years".

...
Further optional arguments that will be passed to fun

Value

(Duration) The age as a duration.

Examples

dob2age(lubridate::as_date("1990-01-01"))

#> [1] "972604800s (~30.82 years)"

dob2age(
  dob = c(
    lubridate::as_date("1990-01-01"),
    lubridate::as_date("1994-01-01"),
    lubridate::as_date("1998-01-01")
  ),
  age_on = lubridate::as_date("2002-12-31"),
  fun = floor
)

#> [1] "378691200s (~12 years)" "252460800s (~8 years)"  "126230400s (~4 years)"