eGFR 2012 CKD-EPI creatinine-cystatin C equation

Using KDIGO 2012 Clinical Practice Guideline for the Evaluation and Management of Chronic Kidney Disease Volume 3 | Issue 1 | January 2013

eGFR_adult_SCr_SCysC(...)

# S3 method for class 'data.frame'
eGFR_adult_SCr_SCysC(.data, SCr, SCysC, Age, male, black, ...)

# S3 method for class 'units'
eGFR_adult_SCr_SCysC(SCr, SCysC, Age, male, black, ...)

# S3 method for class 'numeric'
eGFR_adult_SCr_SCysC(SCr, SCysC, Age, male, black, ...)

Arguments

...

Further optional arguments

.data

(data.frame) A data.frame, optional

SCr

Serum creatinine column name, or vector of units or numeric if .data is not provided

SCysC

Serum Cystatin C column name, or vector of units or numeric if .data is not provided

Age

Age of patient column name, or vector of units or numeric if .data is not provided

male

Male or not column name, or vector of logical (TRUE/FALSE) if .data is not provided

black

Black race or not column name, or vector of logical (TRUE/FALSE) if .data is not provided

Value

Estimated GFR of the same type provided (numeric or units)

Details

See https://kdigo.org/guidelines/ckd-evaluation-and-management/ for more details

Examples

eGFR_adult_SCr_SCysC(eGFR_pt_data,
  SCr = "SCr_", SCysC = "SCysC_",
  Age = "Age_", male = "male_", black = "black_"
)
#> Units: [mL/1.73m2/min]
#>  [1]        NA        NA 166.81886        NA        NA 171.35616        NA
#>  [8]        NA  96.41798        NA        NA  99.04045        NA        NA
#> [15]  97.06854        NA        NA  99.70870        NA        NA  56.10368
#> [22]        NA        NA  57.62964        NA        NA 178.86404        NA
#> [29]        NA 183.72895        NA        NA 103.37985        NA        NA
#> [36] 106.19167        NA        NA 116.50660        NA        NA 119.67546
#> [43]        NA        NA  67.33849        NA        NA  69.17003        NA
#> [50]        NA        NA

eGFR_pt_data %>%
  dplyr::mutate(eGFR = eGFR_adult_SCr_SCysC(
    SCr = SCr_, SCysC = SCysC_,
    Age = Age_, male = male_, black = black_
  ))
#> # A tibble: 51 × 11
#>       SCr_ SCysC_    Age_ male_ black_ height_    BUN_ eGFR_calc_type_     eGFR_
#>    [mg/dl] [mg/L] [years] <lgl> <lgl>      [m] [mg/dl] <chr>               [mL/…
#>  1     0.5   NA        20 FALSE FALSE       NA      NA eGFR_adult_SCr      139. 
#>  2    NA      0.4      20 FALSE FALSE       NA      NA eGFR_adult_SCysC    162. 
#>  3     0.5    0.4      20 FALSE FALSE       NA      NA eGFR_adult_SCr_SCy… 167. 
#>  4     0.5   NA        30 FALSE TRUE        NA      NA eGFR_adult_SCr      150. 
#>  5    NA      0.4      30 FALSE TRUE        NA      NA eGFR_adult_SCysC    155. 
#>  6     0.5    0.4      30 FALSE TRUE        NA      NA eGFR_adult_SCr_SCy… 171. 
#>  7     0.5   NA        20 FALSE FALSE       NA      NA eGFR_adult_SCr      139. 
#>  8    NA      1.2      20 FALSE FALSE       NA      NA eGFR_adult_SCysC     66.8
#>  9     0.5    1.2      20 FALSE FALSE       NA      NA eGFR_adult_SCr_SCy…  96.4
#> 10     0.5   NA        30 FALSE TRUE        NA      NA eGFR_adult_SCr      150. 
#> # ℹ 41 more rows
#> # ℹ 2 more variables: pediatric_ <lgl>, eGFR [mL/1.73m2/min]