eGFR Pediatric SCr, Height and BUN

Using KDIGO 2012 Clinical Practice Guideline for the Evaluation and Management of Chronic Kidney Disease Volume 3 | Issue 1 | January 2013

eGFR_child_SCr_BUN(...)

# S3 method for data.frame
eGFR_child_SCr_BUN(.data, SCr, height, BUN, ...)

# S3 method for units
eGFR_child_SCr_BUN(SCr, height, BUN, ...)

# S3 method for numeric
eGFR_child_SCr_BUN(SCr, height, BUN, ...)

Arguments

...

Further optional arguments

.data

(data.frame) A data.frame, optional

SCr

Serum creatinine column name, or vector of units or numeric if .data is not provided

height

Height of patient column name, or vector of units or numeric if .data is not provided

BUN

Blood urea nitrogen column name, or vector of units or numeric if .data is not provided

Value

Estimated GFR of the same type provided (numeric or units)

Details

See https://kdigo.org/guidelines/ckd-evaluation-and-management/ for more details

Examples

eGFR_child_SCr_BUN(eGFR_pt_data,
  SCr = "SCr_", height = "height_", BUN = "BUN_",
)
#> Units: [mL/1.73m2/min]
#>  [1]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [9]       NA       NA       NA       NA       NA       NA       NA       NA
#> [17]       NA       NA       NA       NA       NA       NA       NA       NA
#> [25]       NA       NA       NA       NA       NA       NA       NA       NA
#> [33]       NA       NA       NA       NA       NA       NA       NA       NA
#> [41]       NA       NA       NA       NA       NA       NA       NA       NA
#> [49]       NA 148.2122       NA

eGFR_pt_data %>%
  dplyr::mutate(eGFR = eGFR_child_SCr_BUN(
    SCr = SCr_, height = height_, BUN = BUN_,
  ))
#> # A tibble: 51 × 11
#>       SCr_ SCysC_    Age_ male_ black_ height_  BUN_ eGFR_…¹ eGFR_ pedia…²  eGFR
#>    [mg/dl] [mg/L] [years] <lgl> <lgl>      [m] [mg/… <chr>   [mL/… <lgl>   [mL/…
#>  1     0.5   NA        20 FALSE FALSE       NA    NA eGFR_a… 139.  FALSE      NA
#>  2    NA      0.4      20 FALSE FALSE       NA    NA eGFR_a… 162.  FALSE      NA
#>  3     0.5    0.4      20 FALSE FALSE       NA    NA eGFR_a… 167.  FALSE      NA
#>  4     0.5   NA        30 FALSE TRUE        NA    NA eGFR_a… 150.  FALSE      NA
#>  5    NA      0.4      30 FALSE TRUE        NA    NA eGFR_a… 155.  FALSE      NA
#>  6     0.5    0.4      30 FALSE TRUE        NA    NA eGFR_a… 171.  FALSE      NA
#>  7     0.5   NA        20 FALSE FALSE       NA    NA eGFR_a… 139.  FALSE      NA
#>  8    NA      1.2      20 FALSE FALSE       NA    NA eGFR_a…  66.8 FALSE      NA
#>  9     0.5    1.2      20 FALSE FALSE       NA    NA eGFR_a…  96.4 FALSE      NA
#> 10     0.5   NA        30 FALSE TRUE        NA    NA eGFR_a… 150.  FALSE      NA
#> # … with 41 more rows, and abbreviated variable names ¹​eGFR_calc_type_,
#> #   ²​pediatric_