eGFR Pediatric SCysC

Using KDIGO 2012 Clinical Practice Guideline for the Evaluation and Management of Chronic Kidney Disease Volume 3 | Issue 1 | January 2013

Usage

eGFR_child_SCysC(...)

# S3 method for class 'data.frame'
eGFR_child_SCysC(.data, SCysC, ...)

# S3 method for class 'units'
eGFR_child_SCysC(SCysC, ...)

# S3 method for class 'numeric'
eGFR_child_SCysC(SCysC, ...)

Arguments

...

Further optional arguments

.data

(data.frame) A data.frame, optional

SCysC

Serum Cystatin C column name, or vector of units or numeric if .data is not provided

Value

Estimated GFR of the same type provided (numeric or units)

Details

See https://kdigo.org/guidelines/ckd-evaluation-and-management/ for more details

Examples

eGFR_child_SCysC(eGFR_pt_data,
  SCysC = "SCysC_"
)
#> Units: [mL/(min*1.73m^2)]
#>  [1]        NA 165.89761 165.89761        NA 165.89761 165.89761        NA
#>  [8]  59.65409  59.65409        NA  59.65409  59.65409        NA 165.89761
#> [15] 165.89761        NA 165.89761 165.89761        NA  59.65409  59.65409
#> [22]        NA  59.65409  59.65409        NA 165.89761 165.89761        NA
#> [29] 165.89761 165.89761        NA  59.65409  59.65409        NA  59.65409
#> [36]  59.65409        NA 165.89761 165.89761        NA 165.89761 165.89761
#> [43]        NA  59.65409  59.65409        NA  59.65409  59.65409        NA
#> [50]        NA 165.89761

eGFR_pt_data %>%
  dplyr::mutate(eGFR = eGFR_child_SCysC(
    SCysC = SCysC_
  ))
#> # A tibble: 51 × 11
#>       SCr_ SCysC_    Age_ male_ black_ height_    BUN_ eGFR_calc_type_     eGFR_
#>    [mg/dl] [mg/L] [years] <lgl> <lgl>      [m] [mg/dl] <chr>               [mL/…
#>  1     0.5   NA        20 FALSE FALSE       NA      NA eGFR_adult_SCr      139. 
#>  2    NA      0.4      20 FALSE FALSE       NA      NA eGFR_adult_SCysC    162. 
#>  3     0.5    0.4      20 FALSE FALSE       NA      NA eGFR_adult_SCr_SCy… 167. 
#>  4     0.5   NA        30 FALSE TRUE        NA      NA eGFR_adult_SCr      150. 
#>  5    NA      0.4      30 FALSE TRUE        NA      NA eGFR_adult_SCysC    155. 
#>  6     0.5    0.4      30 FALSE TRUE        NA      NA eGFR_adult_SCr_SCy… 171. 
#>  7     0.5   NA        20 FALSE FALSE       NA      NA eGFR_adult_SCr      139. 
#>  8    NA      1.2      20 FALSE FALSE       NA      NA eGFR_adult_SCysC     66.8
#>  9     0.5    1.2      20 FALSE FALSE       NA      NA eGFR_adult_SCr_SCy…  96.4
#> 10     0.5   NA        30 FALSE TRUE        NA      NA eGFR_adult_SCr      150. 
#> # ℹ 41 more rows
#> # ℹ 2 more variables: pediatric_ <lgl>, eGFR [mL/(min*1.73m^2)]