Diagnosis of anemia from Hb concentration

KDIGO Clinical Practice Guideline for Anemia in Chronic Kidney Disease Volume 2 | Issue 4 | August (2) 2012

anemia(...)

# S3 method for class 'data.frame'
anemia(.data, Hb, age, male, ...)

# S3 method for class 'units'
anemia(Hb, age, male, ...)

# S3 method for class 'numeric'
anemia(Hb, age, male, ...)

Arguments

...

Further optional arguments

.data

(data.frame) A data.frame, optional

Hb

Hemoglobin concentration column name, or vector of units or numeric if .data is not provided

age

Age of patient column name, or vector of units or numeric if .data is not provided

male

Male or not column name, or vector of logical (TRUE/FALSE) if .data is not provided

Value

Anemia as logical TRUE or FALSE

Details

Adults and children >15 years with CKD when the Hb concentration is <13.0 g/dl (<130 g/l) in males and <12.0 g/dl (<120 g/l) in females.

Children with CKD if Hb concentration is <11.0 g/dl (<110 g/l) in children 0.5-5 years, <11.5 g/dl (115 g/l) in children 5-12 years, and <12.0 g/dl (120 g/l) in children 12-15 years.

See https://kdigo.org/guidelines/anemia-in-ckd/ for more details.

Examples

anemia(anemia_pt_data, Hb = "Hb", age = "age", male = "male")
#>  [1]  TRUE FALSE  TRUE FALSE  TRUE FALSE  TRUE FALSE  TRUE FALSE

anemia_pt_data %>%
  dplyr::mutate(anemia = anemia(Hb = Hb, age = age, male = male))
#> # A tibble: 10 × 4
#>        Hb     age male  anemia
#>    [g/dl] [years] <lgl> <lgl> 
#>  1   12.5      20 TRUE  TRUE  
#>  2   13.5      20 TRUE  FALSE 
#>  3   11.5      20 FALSE TRUE  
#>  4   12.5      20 FALSE FALSE 
#>  5   11.5      13 TRUE  TRUE  
#>  6   12.5      13 TRUE  FALSE 
#>  7   11         7 TRUE  TRUE  
#>  8   12         7 TRUE  FALSE 
#>  9   10.5       3 TRUE  TRUE  
#> 10   11.5       3 TRUE  FALSE